♻️ Extracted route options for Oauth
